common opds start page mixin. emtpy group objects on start page
