+++ codex/views/opds/v1/feed.py
@@ -1,6 +1,6 @@
 """OPDS v1 feed."""
 
-from collections.abc import MutableMapping, Sequence
+from collections.abc import Sequence
 from typing import TYPE_CHECKING, Any, override
 
 from drf_spectacular.utils import extend_schema
@@ -15,7 +15,8 @@ from codex.serializers.browser.settings import OPDSSettingsSerializer
 from codex.serializers.opds.v1 import OPDS1TemplateSerializer
 from codex.settings import BROWSER_MAX_OBJ_PER_PAGE, FALSY
 from codex.version import VERSION
-from codex.views.opds.const import BLANK_TITLE, DEFAULT_PARAMS
+from codex.views.opds.const import BLANK_TITLE
+from codex.views.opds.start import OPDSStartViewMixin
 from codex.views.opds.v1.const import OPDS1EntryData, OpdsNs, RootTopLinks
 from codex.views.opds.v1.entry.entry import OPDS1Entry
 from codex.views.opds.v1.links import OPDS1LinksView
@@ -165,14 +166,8 @@ class OPDS1FeedView(OPDS1LinksView):
         return Response(serializer.data, content_type=self.content_type)
 
 
-class OPDS1StartView(OPDS1FeedView):
+class OPDS1StartView(OPDSStartViewMixin, OPDS1FeedView):
     """OPDS v1 Start Page."""
-
-    IS_START_PAGE = True
-
-    @override
-    def init_params(self) -> MutableMapping[str, Any]:
-        return dict(DEFAULT_PARAMS)
 
     @override
     @extend_schema(
